refactor: restructured codebase
- now it's a base dotnet project instead of visual studio project
