--- FbsDumper.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.14.36017.23 d17.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FbsDumper", "FbsDumper\FbsDumper.csproj", "{D1BFFED4-D603-4938-A0FC-C4C2CA058A13}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D1BFFED4-D603-4938-A0FC-C4C2CA058A13}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D1BFFED4-D603-4938-A0FC-C4C2CA058A13}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D1BFFED4-D603-4938-A0FC-C4C2CA058A13}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D1BFFED4-D603-4938-A0FC-C4C2CA058A13}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C53620CC-E264-4522-AE52-214C285CBAFA}
-	EndGlobalSection
-EndGlobal
